Exclusion (again) of connect on comps.
