+++ src/redux/operators/userOperator.js
@@ -1,9 +1,10 @@
 import { userActions } from "redux/reducersAndActions/userReducer";
+import { userStoreView } from "redux/storeViews/userStoreView";
 
 let userOperator = null;
 /**
  * Get the operator being used to update the store
- * @param {object} store Store object with the redux data
+ * @param {object} dispatch Dispatch function to redux
  */
 export const getUserOperator = (store) => {
     if (!userOperator) userOperator = new UserOperator(store);
@@ -22,6 +23,9 @@ class UserOperator {
     if (!name) {
       return false;
     }
+    /** STUB: Shows how to read data outside components */
+    const loggedName = userStoreView.username(this.store.getState());
+    console.log(loggedName);
     this.store.dispatch(userActions.setUsername(name));
   }
 }
